refactor(pricing-page): rename FAQ components for better scoping
Rename FaqList to PricingPage::FaqList and FaqItem to FaqCard to 
improve component naming consistency and module scoping. Update 
all imports and usage sites accordingly, including templates. 
This change clarifies component ownership and avoids naming 
conflicts within the app.
