+++ app/components/faq-card.hbs
