+++ app/components/pricing-page/faq-list.hbs
@@ -1,7 +1,7 @@
 <div class="w-full max-w-3xl mx-auto" ...attributes>
   <div class="flex flex-col gap-2.5">
     {{#each this.faqs as |faq|}}
-      <FaqItem @faq={{faq}} @onToggle={{this.handleToggle}} @isOpen={{eq faq this.openItem}} />
+      <FaqCard @faq={{faq}} @onToggle={{this.handleToggle}} @isOpen={{eq faq this.openItem}} />
     {{/each}}
   </div>
 </div>
