+++ app/templates/pay.hbs
@@ -86,7 +86,7 @@
       Frequently Asked Questions
     </h2>
 
-    <FaqList class="mt-8" />
+    <PricingPage::FaqList class="mt-8" />
 
     <h2 class="text-2xl font-bold text-gray-700 md:text-3xl max-w-3xl text-center text-pretty mt-16">
       Join over 200k of the world’s best engineers:
